python/requirements.txt
numpy>=1.17
scipy>=1.3.1
scikit-learn>=0.22
numba>=0.51.2
pynndescent>=0.5
tbb>=2019.0
tqdm
umap>=0.5
pillow>=8.3